[BUGFIX] Replace array constant for compatibility with PHP 5.5
As of PHP version 5.6 it is possible to define class constants as arrays.
In PHP versions < 5.6 this results in a fatal PHP error ("Arrays are not
allowed in class constants").

In order to make EXT:nagios version 2.x compatible with PHP version 5.5,
these constants have been replaced with normal variables.

Resolves: issue #5 (GitHub)