[!!!][BUGFIX] Add missing keyword 'version' to list of extensions
In previous versions, the extension list contains the keyword "version"
between the extension key and the version number. For example:

EXT:realurl-version-2.0.14

This keyword has been accidentally dropped in version 2.x. As a
consequence version 1.0.0.4 of the Nagios check_typo3 plugin can be used,
which processes the extension list generated by TYPO3 CMS with and
without the keyword.

This patch re-introduces the keyword "version" in the extension list.

BREAKING CHANGE: in cases where the Nagios plugin has been updated
manually to ignore the keyword, the new version of EXT:nagios possibly
breaks the plugin.

References:
https://github.com/schams-net/nagios/issues/6
https://exchange.nagios.org/directory/Plugins

Fixes #6