[BUGFIX] Use fallback method if ext version extraction fails
In a classic TYPO3 installation without Composer, the following function
returns an empty string for extensions that are not installed/loaded:
ExtensionManagementUtility::getExtensionVersion()

This results in an extension list showing disabled extensions without
version string (only in classic installations).

The fix implements a fallback method for these cases and checks if the
determined version string is syntactically valid. Extensions with invalid
versions are now excluded from the extension list.